Add compatibility for solitude
