+++ webdriverPoller.py
@@ -1,3 +1,9 @@
+"""
+Use: python webdriverPoller.py --config <configFile>
+defaults to pollingConfig.json
+"""
+
+from argparse import ArgumentParser
 from selenium import webdriver
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.by import By
@@ -8,36 +14,38 @@ from selenium.webdriver.chrome.options import Options
 
 from tempfile import mkdtemp
 import time
-import os
 import datetime as dt
 
-LOGIN_URL = "https://reservenski.parksolitude.com/login"
+from pollingConfig import ReservationType, load_json_config
 
 
-class SolitudeParking:
-    def __init__(self, driver, credentials):
+class WebdriverPoller:
+    def __init__(self, driver, credentials, resort):
         self.driver = driver
-        self.driver.get(LOGIN_URL)
         self.username = credentials['username']
         self.password = credentials['password']
+        self.resort = resort.value
+        self.login_url = resort.value.base_url + "/login"
+        self.parking_codes_url = resort.value.base_url + "/parking-codes"
+        self.calendar_url = resort.value.base_url + "/select-parking"
 
     def start_session(self):
         """Creates the selenium session
         """
         # Setup option
         options = Options()
-        if self.headless:
-            options.add_argument("--headless=new")
 
         # Open
         self.driver = webdriver.Chrome(options=options)
-        self.driver.get(LOGIN_URL)
+        self.driver.get(self.login_url)
 
     def login(self):
         """Login the user with enviornment variables
         """
+        self.driver.get(self.login_url)
+
         # Get elements
-        WebDriverWait(self.driver, 15).until(EC.presence_of_all_elements_located(
+        WebDriverWait(self.driver, 5).until(EC.presence_of_all_elements_located(
             (By.ID,
              'emailAddress')
         ))
@@ -56,9 +64,9 @@ class SolitudeParking:
         submit_btn.click()
 
         # Wait until finished and re-directed
-        WebDriverWait(self.driver, 15).until(EC.presence_of_all_elements_located(
+        WebDriverWait(self.driver, 5).until(EC.presence_of_all_elements_located(
             (By.XPATH,
-             '//div[text()="Reserve Parking Before Arriving at Solitude"]')
+             f'//div[text()="{self.resort.main_screen_string}"]')
         ))
 
     def activate_code(self):
@@ -66,11 +74,11 @@ class SolitudeParking:
         this will give us the $0 option in booking
         """
         # Go to codes
-        self.driver.get("https://reservenski.parksolitude.com/parking-codes")
+        self.driver.get(self.parking_codes_url)
 
         # Wait for button
         btn_xpath = '//button[text()="Reserve Parking"]'
-        WebDriverWait(self.driver, 15).until(EC.presence_of_element_located(
+        WebDriverWait(self.driver, 5).until(EC.presence_of_element_located(
             (By.XPATH, btn_xpath)))
 
         # Select Code
@@ -80,16 +88,15 @@ class SolitudeParking:
 
         # This will automatically navigate to calendar, just don't let
         # go until this is complete
-        WebDriverWait(self.driver, 15).until(EC.presence_of_all_elements_located(
+        WebDriverWait(self.driver, 5).until(EC.presence_of_all_elements_located(
             (By.XPATH, '//div[text()="What Day Are You Parking at "]')
         ))
 
     def go_to_selection_calendar(self):
         """Goes to the calendar selection page, if not already there
         """
-        if (self.driver.current_url != "https://reservenski.parksolitude.com/select-parking"):
-            self.driver.get(
-                "https://reservenski.parksolitude.com/select-parking")
+        if (self.driver.current_url != self.calendar_url):
+            self.driver.get(self.calendar_url)
 
     def navigate_to_date(self, date):
         """On the calendar selection page, ensure that the month of interest
@@ -99,7 +106,7 @@ class SolitudeParking:
             date (datetime): date we want to be able to select
         """
         # Wait for calendar to load
-        WebDriverWait(self.driver, 15).until(EC.presence_of_element_located(
+        WebDriverWait(self.driver, 5).until(EC.presence_of_element_located(
             (By.CLASS_NAME, 'ParkingSelection_calendarWrapper__kS9AJ')))
 
         # Navigate to appropriate month
@@ -146,42 +153,57 @@ class SolitudeParking:
         # page (to support the animation of switching months), so some dates
         # are duplicated. We find the first one that is enabled and click it.
         btns = self.driver.find_elements(
-            By.XPATH, '//div[@aria-label="{}"]'.format(date_string))
+            By.XPATH, '//div[@aria-label="{}"]'.format(date_string)) + self.driver.find_elements(
+            By.XPATH, '//div[@aria-label="Today, {}"]'.format(date_string))
+
+        if len(btns) == 0:
+            raise Exception(
+                "Can't find date button for {}".format(date_string))
+
         for btn in btns:
-            if btn.is_enabled() and btn.is_displayed():
+            if btn.is_enabled() and btn.is_displayed() and btn.location['x'] > 0 and btn.location['x'] < 1696 and btn.location['y'] > 0:
                 btn.click()
                 break
 
-    def select_parking_option(self):
+    def select_parking_option(self, reservation_type: ReservationType):
         """Clicks the Season Pass Holder option once the date
         is selected, which redirect to purchasing
         """
         # Wait for button to load and click
-        btn_xpath = '//div[text()="Season Pass Holders"]/parent::*/parent::*'
-        WebDriverWait(self.driver, 3).until(EC.presence_of_element_located(
+        if (reservation_type == ReservationType.PASSHOLDER):
+            btn_xpath = f'//div[text()="{self.resort.passholder_string}"]/parent::*/following-sibling::div[text()="$0"]/parent::*'
+        elif (reservation_type == ReservationType.CARPOOL):
+            btn_xpath = f'//div[text()="{self.resort.carpool_string}"]/parent::*/following-sibling::div[text()="$0"]/parent::*'
+
+        WebDriverWait(self.driver, 1).until(EC.presence_of_element_located(
             (By.XPATH, btn_xpath)))
         btn = self.driver.find_element(
             By.XPATH, btn_xpath)
         btn.click()
 
         # Wait for honk to load
-        WebDriverWait(self.driver, 15).until(EC.presence_of_element_located(
+        WebDriverWait(self.driver, 5).until(EC.presence_of_element_located(
             (By.XPATH, '//div[text()="Checkout"]')))
 
-    def reserve(self):
+    def reserve(self, reservation_type: ReservationType):
         """Once we are on honk, pull trigger and reserve the pass!
         """
+        if (reservation_type == ReservationType.PASSHOLDER):
+            btn_xpath = '//div[text()="Redeem A Pass"]/parent::*/parent::*'
+        elif (reservation_type == ReservationType.CARPOOL):
+            btn_xpath = '//div[text()="Park For Free"]/parent::*/parent::*'
+
         # Ensure that the button is there
-        WebDriverWait(self.driver, 15).until(EC.presence_of_element_located(
-            (By.XPATH, '//div[text()="Redeem A Pass"]')))
+        WebDriverWait(self.driver, 5).until(EC.presence_of_element_located(
+            (By.XPATH, btn_xpath)))
 
         # Get Button
         btn = self.driver.find_element(
-            By.XPATH, '//div[text()="Redeem A Pass"]/parent::*/parent::*')
+            By.XPATH, btn_xpath)
         btn.click()
 
         # Wait for Confirm button
-        WebDriverWait(self.driver, 15).until(EC.presence_of_element_located(
+        WebDriverWait(self.driver, 5).until(EC.presence_of_element_located(
             (By.XPATH, '//button[text()="Confirm"]')))
 
         # Click confirm button
@@ -203,33 +225,65 @@ class SolitudeParking:
         self.select_parking_option()
         self.reserve()
 
+    def poll_for_reservation(self, target_date, reservation_type: ReservationType):
+        """Polls for a reservation for the given date"""
+
+        self.login()
+        if (reservation_type == ReservationType.PASSHOLDER):
+            self.activate_code()
+        self.go_to_selection_calendar()
+        self.navigate_to_date(target_date)
+
+        # Refresh date is day after target date, unless target date is the last
+        # day of the month, in which case we refresh to the previous day.
+        refresh_date = target_date + dt.timedelta(days=1)
+        if refresh_date.month != target_date.month:
+            refresh_date = target_date - dt.timedelta(days=1)
+
+        # Keep trying to select the parking option until it is available
+        has_availability = False
+        while not has_availability:
+            try:
+                self.select_parking_option(reservation_type)
+                has_availability = True
+            except Exception as e:
+                # Option is still sold out, navigate to next/previous day and try again
+                time.sleep(5)
+                self.navigate_to_date(refresh_date)
+
+                time.sleep(1)
+                self.navigate_to_date(target_date)
+
+        self.reserve(reservation_type)
+        print("Reservation secured")
+
 
 if __name__ == "__main__":
+    parser = ArgumentParser()
+    parser.add_argument("--config", default="pollingConfig.json")
+    args = parser.parse_args()
+
+    config = load_json_config(args.config)
+
     options = webdriver.ChromeOptions()
 
     options.add_argument("--headless=new")
     options.add_argument('--no-sandbox')
     options.add_argument("--disable-gpu")
     options.add_argument("--window-size=1280x1696")
-    # options.add_argument("--single-process")
-    # options.add_argument("--disable-dev-shm-usage")
-    # options.add_argument("--disable-dev-tools")
-    # options.add_argument("--no-zygote")
-    # options.add_argument(f"--user-data-dir={mkdtemp()}")
-    # options.add_argument(f"--data-path={mkdtemp()}")
-    # options.add_argument(f"--disk-cache-dir={mkdtemp()}")
-    # options.add_argument("--remote-debugging-port=9222")
     chrome = webdriver.Chrome(options=options)
-    sp = SolitudeParking(chrome)
-    # sp.start_session()
-    sp.login()
-    # sp.activate_code()
-    # sp.go_to_selection_calendar()
-    # sp.navigate_to_date(dt.datetime(2024, 2, 19))
-    # sp.select_parking_option()
 
-#     # sp.reserve()
-
-#     # just keep page open for debuggin
-#     while True:
-#         pass
+    sp = WebdriverPoller(chrome, credentials={
+        "username": config.username,
+        "password": config.password
+    },
+        resort=config.resort
+    )
+    sp.start_session()
+    target = config.dates[0]
+    sp.poll_for_reservation(
+        dt.datetime(target.year, target.month, target.day),
+        config.reservation_type)
+    # sp.reserve()
+    # Set a breakpoint here to keep page open
+    print("Done")
